+++ CHANGELOG.md
@@ -2,6 +2,7 @@ Pocket Casts Change Log
 ======================
 
   * Switched to Chromium backend for audio playback without Flash plugin.
+  * The Beta version of the new player was integrated.
 
 1.0 - September 17th, 2017
 -----------------------